+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed reportsV2: map display, settings, styles
 - fixed reportsV2: Incorrect display of the graph and non-working control buttons 
 - fixed resizing movable visualisation
+- fixed change osmserver in map settings
 
 ## [1.85.0]